moved all nutrition related components to screens folder
